requirements.txt
pip>20
flask
wheel
colorlog
flask_testing
pyhamcrest
requests
nose
selenium

